methods

+++ cars/views.py
@@ -10,6 +10,8 @@ from cars import db_communication as db
 @csrf_exempt
 def add_car(request: HttpRequest):
     try:
+        if request.method != 'POST':
+            return HttpResponseBadRequest("Wrong request method (GET, POST, PUT, DELETE)")
         values = json.loads(request.body)
         token = request.headers.get('Authorization')
         db.add_car(token, mark=values["mark"], model=values["model"], color=values["color"],
@@ -28,6 +30,8 @@ def add_car(request: HttpRequest):
 @csrf_exempt
 def delete_car(request: HttpRequest, id_: int):
     try:
+        if request.method != 'DELETE':
+            return HttpResponseBadRequest("Wrong request method (GET, POST, PUT, DELETE)")
         token = request.headers.get('Authorization')
         status = db.delete_car(token, id_)
         if not status:
